requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
passlib[bcrypt]
python-jose[cryptography]
pydantic
fastapi-mail
python-dotenv
httpx
